+++ example/lib/main.dart
@@ -24,7 +24,11 @@ class _MyHomePageState extends State<MyHomePage> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(title: Text('bottom_bar_with_sheet v0.3.2')),
+      appBar: AppBar(
+        title: Text('bottom_bar_with_sheet v0.4.0',
+            style: TextStyle().copyWith(color: Colors.white)),
+        backgroundColor: const Color(0xFF2B65E3),
+      ),
       body: Center(child: Text("Place for your content")),
       bottomNavigationBar: BottomBarWithSheet(
         selectedIndex: _selectedIndex,