+++ lib/src/builders.dart
@@ -1,0 +1,3 @@
+import 'package:flutter/material.dart';
+
+typedef FullBuilder = Widget Function(BuildContext context);