+++ lib/src/positions.dart
@@ -1,2 +1,2 @@
-enum MainButtonPosition { Left, Middle, Right }
+enum MainButtonPosition { Left, Middle, Right, Outside }
 enum RowPosition { Left, Right }